Update: Move common functions to utils
